Journal/requirement.txt

fastapi             
uvicorn[standard]  
sqlalchemy          
pydantic           
streamlit          
pandas              
requests            
feedparser          
beautifulsoup4      
python-dotenv       